==> notebooks/1_data_preprocessing.ipynb <==
import numpy as np
import pandas as pd
import matplotlib
from matplotlib import pyplot as plt
import os
import re
import copy
from sklearn.decomposition import PCA
import warnings; warnings.simplefilter('ignore')
%matplotlib inline

path1 = os.path.join(os.getcwd(), '1/')
path2 = os.path.join(os.getcwd(), '2/')

## Helper Functions

def get_gene_set(mat):
    return set(mat.keys())

def remove_repeated_genes(mat, repeated_genes_dict):
    for g, c in repeated_genes_dict.items():
        gene_df = mat[g]
        mean = pd.DataFrame.sum(gene_df, axis=1)/float(c)
        # drop the repetitions
        rep_idx = np.where(mat.columns.get_loc(g) == True)[0]
        mat.drop(mat.columns[rep_idx], axis=1, inplace=True)
        # insert the averaged col
        mat[g] = copy.deepcopy(mean)
    return mat

def get_intersection_gene_df(mat, intersection_gene_list):
    return mat[intersection_gene_list]

## Dataset 1

mat1 = pd.read_table(os.path.join(path1, 'without_header.txt'))

mat1.head(2)

mat1.shape

mat1_rpkm_only = mat1.iloc[:, : int(7030/2)]
mat1_rpkm_only.head(2)

mat1_rpkm_only.shape  

header = pd.read_table(os.path.join(path1, 'header.txt'))

header.head()

column_headers = {}
for old, new in zip(mat1_rpkm_only.keys(), header.keys()):
    column_headers[old] = new

mat1_rpkm_only.rename(columns=column_headers, inplace=True)
mat1_rpkm_only.head(2)

#Drop the HP1502401_N13 column
mat1_rpkm_only.drop('HP1502401_N13', axis=1, inplace=True)
mat1_rpkm_only.head(2)

mat1_rpkm_only.tail(2)

# From README, 26180:26271 correspond to data for the 92 external RNA spike-in controls (ERCCs). Deleting these
mat1_rpkm_only.drop(mat1_rpkm_only.tail(93).index,inplace=True)
mat1_rpkm_only.tail()

# Delete all columns with only 0s
mat1_rpkm_only = mat1_rpkm_only.loc[:, (mat1_rpkm_only != 0).any(axis=0)]
# Delete all rows with only 0s
mat1_rpkm_only = mat1_rpkm_only.loc[(mat1_rpkm_only != 0).any(axis=1), :]

mat1_rpkm_only.shape

mat1_rpkm_only.set_index('#samples',inplace=True)

mat1_rpkm_only.head(2)

mat1T = mat1_rpkm_only.T
mat1T.head(2)

# compare number of unique genes with total cols
genes1 = get_gene_set(mat1T)
print('# of unique genes = %s' %(len(genes1)))
print('# of cols = %s' %(len(mat1T.keys())))

# Merging same gene cols by taking mean 
gene_occ_dict = {}
for g in mat1T.keys():
    if g not in gene_occ_dict.keys():
        gene_occ_dict[g] = 1
    else:
        gene_occ_dict[g] += 1

repeated_genes_dict = {}
for g, c in gene_occ_dict.items():
    if c > 1:
        repeated_genes_dict[g] = c

mat1T = remove_repeated_genes(mat1T, repeated_genes_dict)
mat1T.shape

## Dataset 2

mat2 = pd.read_table(os.path.join(path2, 'GSE81608_human_islets_rpkm.txt'))

mat2.head(3)

mat2['gene.id'] = mat2['gene.id'].astype(int)

### Mapping from entrez id (gene.id) to symbols

temp = mat2['gene.id'].astype(str)
with open(os.path.join(path2, 'entrez.txt'), 'w') as f:
    for i in temp.values:
        f.write(str(i) + "\n")

e2s = {}
with open(os.path.join(path2, 'symbols.txt') ,'rb') as f: # obtained by running an R script
    for entrez, sym in zip(temp, f.readlines()):
        e2s[int(entrez)] = str(sym)[2:-3]

mat2.insert(loc=1, column='gene_symbol', value=str)
mat2.head(2)

for i in mat2.index:
    gene_id = mat2.iloc[i,0]
    mat2.iloc[i,1] = e2s[gene_id]

mat2.head(2)

mat2.drop('gene.id', axis=1, inplace=True)
mat2.head(2)

mat2.set_index('gene_symbol', inplace=True)

mat2.head(2)

mat2T = mat2.T
mat2T.head(2)

# Dropping NA gene columns
drop_idx = np.where(mat2T.columns.values =='NA')[0]

len(drop_idx)

mat2T.drop(mat2T.columns[drop_idx], axis=1, inplace=True)
mat2T.head(2)

mat2T.shape

# Delete all columns with only 0s
mat2T = mat2T.loc[:, (mat2T != 0).any(axis=0)]
# Delete all rows with only 0s
mat2T = mat2T.loc[(mat2T != 0).any(axis=1), :]
mat2T.shape

No cells were removed. 38250 - 32453 = 5797 genes were removed

# Checking for repetitive cols
genes2 = get_gene_set(mat2T)
len(genes2)

Clearly no repetitive cols in dataset 2

## Intersection Genes

intersection_genes = genes1.intersection(genes2)

len(intersection_genes)

mat1_intrscn = get_intersection_gene_df(mat1T, list(intersection_genes))
mat2_intrscn = get_intersection_gene_df(mat2T, list(intersection_genes))

print(mat1_intrscn.shape)
print(mat2_intrscn.shape)

mat1_intrscn.to_pickle(os.path.join(path1,'1_mat1.pkl'))
mat2_intrscn.to_pickle(os.path.join(path2,'1_mat2.pkl'))
