==> notebooks/2_phenotype_extraction.ipynb <==
import numpy as np
import pandas as pd
import matplotlib
from matplotlib import pyplot as plt
import os
import re
import copy
from sklearn.decomposition import PCA
import warnings; warnings.simplefilter('ignore')
%matplotlib inline

path1 = os.path.join(os.getcwd(), '1/')
path2 = os.path.join(os.getcwd(), '2/')

## Dataset 1

mat1 = pd.read_pickle(os.path.join(path1, '1_mat1.pkl'))

mat1.head(2)

labels1 = pd.read_table(os.path.join(path1, 'E-MTAB-5061.sdrf.txt'))

labels1.head(3)

np.unique(labels1['Characteristics[cell type]'])

## Dataset 2 

mat2 = pd.read_pickle(os.path.join(path2, '1_mat2.pkl'))

mat2.head(2)

labels2 = pd.read_table(os.path.join(path2, 'GSE81608_series_matrix_edited.txt'))

labels2.head(2)

label_types = ['subtype', 'condition', 'age', 'ethnicity', 'gender']
for idx in labels2.index:
    labels2['!Sample_title'].loc[idx] = label_types[idx]
labels2.head()

labels2.head(2)

labels2.set_index('!Sample_title', inplace=True)

labels2 = labels2.T
labels2.head(2)

# Unique cell types
np.unique(labels2['subtype'])

## Cell Type Distributions
Cells common to both:
    - Alpha
    - Beta
    - Delta
    - PP (aka Gamma) cells
### Dataset 1

cell_types1 = ['alpha cell', 'beta cell', 'delta cell', 'gamma cell']

count1 = 0
for c in cell_types1:
    cell_type_count = len(labels1[labels1['Characteristics[cell type]'] == c])
    count1 += cell_type_count
    print(c, str(cell_type_count))
print("Total usable cells = ", str(count1))

### Dataset 2

cell_types2 = ['cell subtype: PP', 'cell subtype: alpha', 'cell subtype: beta', 'cell subtype: delta']

count2 = 0
for c in cell_types2:
    cell_type_count = len(labels2[labels2['subtype'] == c])
    count2 += cell_type_count
    print(c, str(cell_type_count))
print("Total usable cells = ", str(count2))

## Labels(Note: Gamma and PP cells are same.)
## PP  = Pancreatic Polypetide cell
cell_type_dict_1 = {'alpha cell': 0, 
                    'beta cell': 1,
                    'delta cell': 2,
                    'gamma cell': 3}


cell_type_dict_2 = {'cell subtype: alpha': 0, 
                    'cell subtype: beta': 1,
                    'cell subtype: delta': 2,
                    'cell subtype: PP': 3}

## Disease State
### Dataset 1

np.unique(labels1['Characteristics[disease]'])

### Dataset 2

np.unique(labels2['condition'])

# Labels: T2D is Type II diabetes mellitus
dis_cond_dict_1 = {'normal': 0 , 'type II diabetes mellitus' : 1}
dis_cond_dict_2 = {'condition: non-diabetic': 0 , 'condition: T2D' : 1}

## Adding Phenotype cols
### Dataset 1

mat1['cell_type'] = pd.Series(dtype=int, index=mat1.index)
mat1['disease_condition'] =  pd.Series(dtype=int, index=mat1.index)

for cell_id in mat1.index:
    cell_type = labels1[labels1['Source Name'] == cell_id]['Characteristics[cell type]'].values[0]
    dis_cond =  labels1[labels1['Source Name'] == cell_id]['Characteristics[disease]'].values[0]
    
    if str(cell_type) not in cell_type_dict_1.keys():
        mat1.loc[cell_id]['cell_type'] = -1
    else:
        mat1.loc[cell_id]['cell_type'] = cell_type_dict_1[cell_type]
    
    if str(dis_cond).rstrip() not in dis_cond_dict_1.keys():
        mat1.loc[cell_id]['disease_condition'] = -1
    else:
        mat1.loc[cell_id]['disease_condition'] = dis_cond_dict_1[dis_cond]

mat1.shape

### Dataset 2

mat2['cell_type'] = pd.Series(dtype=int, index=mat2.index)
mat2['disease_condition'] =  pd.Series(dtype=int, index=mat2.index)

labels2_new_idx = []
for i in labels2.index:
    sample_num = i[29:] # cell numbers
    labels2_new_idx.append('Sample_'+str(sample_num))

new_idx = pd.Series(labels2_new_idx, index=labels2.index)

labels2['sample_id'] = new_idx

labels2.head()

labels2.set_index('sample_id', inplace=True)

labels2.head()

for cell_id in mat2.index:
    cell_type = labels2.loc[cell_id]['subtype']
    dis_cond =  labels2.loc[cell_id]['condition']
    if str(cell_type) not in cell_type_dict_2.keys():
        mat2.loc[cell_id]['cell_type'] = -1
    else:
        mat2.loc[cell_id]['cell_type'] = cell_type_dict_2[cell_type]
    
    if str(dis_cond) not in dis_cond_dict_2.keys():
        mat2.loc[cell_id]['disease_condition'] = -1
    else:
        mat2.loc[cell_id]['disease_condition'] = dis_cond_dict_2[dis_cond]

mat2.head()

## Filtering cells based on subtype and disease condition
### Dataset 1

mat1_fil = mat1[ (mat1.disease_condition != -1.0) & (mat1.cell_type != -1.0)]

mat1_fil.shape

mat2_fil = mat2[(mat2.disease_condition != -1.0) & (mat2.cell_type != -1.0)]

mat2_fil.shape

mat1_fil.to_pickle(os.path.join(path1, '2_mat1_with_pheno.pkl'))
mat2_fil.to_pickle(os.path.join(path2, '2_mat2_with_pheno.pkl'))

print(1)
