==> notebooks/6_KMeans.ipynb <==
import warnings; warnings.simplefilter('ignore')
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import os
import copy
from sklearn.preprocessing import scale
from sklearn.decomposition import PCA
from sklearn.manifold import TSNE
from sklearn.model_selection import train_test_split
from sklearn.model_selection import KFold
from sklearn.metrics import precision_recall_fscore_support
from sklearn.neighbors import KNeighborsClassifier

from sklearn.metrics import classification_report

%matplotlib inline
%pylab inline

y.shape

x = pd.read_csv("x_pca", header=None).values
y = pd.read_csv("y_pca", header=None).values

kf = KFold(n_splits=10)

for train_index, test_index in kf.split(x):
    x_train = x[train_index, :]
    x_test = x[test_index, :]
    y_train = y[train_index]
    y_test = y[test_index]
    knn = KNN(x_train, y_train, x_test, y_test)
    
    knn.result_print()
    
    

tsne = pd.read_csv("tsne_pca_with_pheno.csv", header=None)
x = tsne.iloc[:, 0:2].values
Y = tsne.iloc[:, 2].values

tsne.shape

y.shape

## gan

tsne = pd.read_csv("tsne_gans_with_pheno.csv", header=None)
x = tsne.iloc[:, 0:2].values
Y = tsne.iloc[:, -1].values

kf = KFold(n_splits=10)
for train_index, test_index in kf.split(x):
    x_train = x[train_index, :]
    x_test = x[test_index, :]
    y_train = y[train_index]
    y_test = y[test_index]
    knn = KNN(x_train, y_train, x_test, y_test)
    
    knn.result_print()



class KNN:
    def __init__(self, x_train, y_train, x_test, y_test, k=[20]):
        self.x_train = x_train
        self.y_train = y_train
        self.x_test = x_test
        self.y_test = y_test
        self.k = k
        self.name = 'KNN'
        self.results = {}
        
        for k in self.k:
            self.results[k] = {'precision': 0, 'recall': 0, 'f1_score': 0, 'time': 0}
        
        self.driver()

    def knn_classify(self, knn):
        clf = knn.fit(self.x_train, self.y_train)
        predictions = clf.predict(self.x_test)

        return predictions

    def classify(self, k):
        knn = KNeighborsClassifier(n_neighbors=k, algorithm='brute', metric="euclidean")

        predictions = self.knn_classify(knn)
        #precision, recall, f1_score = classification_report(self.y_test, predictions)
        targets = ["alpha", "beta", "delta", "gamma"]
        print("result for k: {}".format(k))
        print(classification_report(self.y_test, predictions, target_names=targets))

        #return precision, recall, f1_score
    
    def driver(self):
        for k in self.k:
            self.classify(k)
            

    def result_print(self):
        pass
        
        
        

classification_report??
